Refactor more top-level files as TypeScript
