+++ rollup.browser-config.js
@@ -5,8 +5,8 @@ import typescript from '@rollup/plugin-typescript'
 export default {
   input: {
     index: 'src/index.ts',
-    types: 'src/types.js',
-    util: 'src/util.js'
+    types: 'src/types.ts',
+    util: 'src/util.ts'
   },
   output: { dir: 'browser/dist', format: 'esm', preserveModules: true },
   plugins: [
