--- src/log.d.ts
@@ -1,4 +1,0 @@
-import type { LogLevelId } from './constants.js'
-
-export function debug(logLevel: LogLevelId, ...messages: any[]): void
-export function warn(logLevel: LogLevelId, warning: string | Error): void
