+++ src/log.ts
@@ -1,13 +1,13 @@
 /* global console, process */
 
-import { LogLevel } from './constants.js'
+import { LogLevel, LogLevelId } from './constants.js'
 
-export function debug(logLevel, ...messages) {
+export function debug(logLevel: LogLevelId, ...messages: any[]) {
   if (LogLevel.indexOf(logLevel) >= LogLevel.DEBUG)
     console.log.apply(console, messages)
 }
 
-export function warn(logLevel, warning) {
+export function warn(logLevel: LogLevelId, warning: string | Error) {
   if (LogLevel.indexOf(logLevel) >= LogLevel.WARN) {
     if (typeof process !== 'undefined' && process.emitWarning)
       process.emitWarning(warning)
