+++ src/stringify/stringifyString.d.ts
@@ -1,0 +1,9 @@
+import { Scalar } from '../ast/index'
+import { Schema } from '../doc/Schema'
+
+export function stringifyString(
+  item: Scalar,
+  ctx: Schema.StringifyContext,
+  onComment?: () => void,
+  onChompKeep?: () => void
+): string
