+++ src/types.ts
@@ -1,11 +1,4 @@
-export {
-  binaryOptions,
-  boolOptions,
-  intOptions,
-  nullOptions,
-  strOptions
-} from './tags/options.js'
-
+export * from './tags/options.js'
 export { Schema } from './doc/Schema.js'
 export {
   Alias,
