+++ src/util.ts
@@ -1,7 +1,5 @@
 export { findPair, toJS } from './ast/index.js'
-
 export { stringifyNumber } from './stringify/stringifyNumber.js'
 export { stringifyString } from './stringify/stringifyString.js'
 export { Type } from './constants.js'
-
-export { YAMLError, YAMLParseError, YAMLWarning } from './errors.js'
+export * from './errors.js'
