+++ types.d.ts
@@ -1,3 +1,1 @@
-export * from './dist/ast/index'
-export { Schema } from './dist/doc/Schema'
-export * from './dist/tags/options'
+export * from './dist/types'
